migrate tooltips to new props
